==> Bowl_Game_Predictions/2018/1-PreProcess.ipynb <==
import pandas as pd
import numpy as np

ind_games = pd.read_csv("data/final_data/all_games_16_18.csv")

ind_games = ind_games.drop('Team2_int_re_number',axis=1)

clean_ind_games = ind_games.drop(['game_id','Team1_team_id','Team2_team_id','Team1','Team2'],axis=1)

### Target split

clean_ind_games['Team2_win'].value_counts()/clean_ind_games.shape[0]

### Re-shuffle dataframe to get equal target variables 

target_yes = clean_ind_games[clean_ind_games['Team2_win'] == 1]
target_no = clean_ind_games[clean_ind_games['Team2_win']==0]

sample_no = int((target_yes.shape[0] - target_no.shape[0])/2)

flip_df = target_yes.sample(n=sample_no)
remove_index = flip_df.index.tolist()

flip_df.head()

orig_cols = flip_df.columns.tolist()

target_yes = target_yes.drop(remove_index)

### Rename and flip values
t1_cols_rename = flip_df[flip_df.columns[flip_df.columns.str.contains('Team1')]].columns.tolist()
sep_df = flip_df[t1_cols_rename]
t1_cols_rename = [word.replace('1','2') for word in t1_cols_rename]
sep_df.columns = t1_cols_rename

t2_cols_rename = flip_df[flip_df.columns[flip_df.columns.str.contains('Team2')]].columns.tolist()
sep_df2 = flip_df[t2_cols_rename]
t2_cols_rename = [word.replace('2','1') for word in t2_cols_rename]
sep_df2.columns = t2_cols_rename

flip_df = pd.concat([sep_df,sep_df2],axis=1)
flip_df = flip_df.drop('Team1_win',axis=1)
flip_df['Team2_win'] = np.where(flip_df['Team2_team_score']>flip_df['Team1_team_score'],1,0)
flip_df = flip_df[orig_cols]

target_yes.shape

target_no = pd.concat([target_no,flip_df],axis=0)

clean_ind_games2 = pd.concat([target_yes,target_no],axis=0).sample(frac=1).reset_index(drop=True)

clean_ind_games2.to_csv('data/final_data/clean_shuffled_games.csv',index=False)

### Create Scale Model 

scale_df = clean_ind_games2.drop(['Team2_win','Team1_team_score','Team2_team_score'],axis=1).convert_objects(convert_numeric=True)
scale_df = scale_df.fillna(0)

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaled_data = scaler.fit_transform(scale_df)

from sklearn.externals import joblib 
scaler_filename = "models/scaler.sav"
joblib.dump(scaler, scaler_filename) 

### Train/Test

from sklearn.model_selection import train_test_split

clean_ind_games2 = clean_ind_games2.drop(['Team1_team_score','Team2_team_score'],axis=1).convert_objects(convert_numeric=True)
clean_ind_games2 = clean_ind_games2.fillna(0)

trainingSet, testSet = train_test_split(clean_ind_games2, test_size=0.3,stratify=None)

trainingSet.to_csv("data/final_data/clean_train.csv",index=False)
testSet.to_csv("data/final_data/clean_test.csv",index=False)

### Prepare Data to search through

search_df = ind_games.drop(['Team1_team_id','Team2_team_id','Team1_team_score','Team2_team_score',
                           'Team2_win'],axis=1)

search_df.shape

game_ids = search_df['game_id']

team1_df = search_df[search_df.columns[search_df.columns.str.contains('Team1')].tolist()]
team2_df = search_df[search_df.columns[search_df.columns.str.contains('Team2')].tolist()]

t1_cols_rename = team2_df.columns
t1_cols_rename = [word.replace('2','1') for word in t1_cols_rename]
team2_df.columns = t1_cols_rename

search_df_fin = pd.concat([pd.concat([game_ids,team1_df],axis=1),pd.concat([game_ids,team2_df],axis=1)],axis=0)

search_df_fin.shape

search_df_fin.to_csv('data/final_data/team_games_data.csv',index=False)
