==> Bowl_Game_Predictions/2018/2-SVM.ipynb <==
import pandas as pd
import utils
from sklearn.feature_selection import SelectFromModel
from sklearn.linear_model import LassoCV

from time import time
import numpy as np
from sklearn import svm
from sklearn.model_selection import RandomizedSearchCV
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_curve, auc

from sklearn.externals import joblib
scale = joblib.load("models/scaler.sav")

### Feature Selection
Lasso 

train = pd.read_csv("data/final_data/clean_train.csv")
X,y,c_names = utils.setup_data(train,scale,'Team2_win',names=True)
train.head()

test = pd.read_csv("data/final_data/clean_test.csv")
X_test,y_test = utils.setup_data(test,scale,target_col='Team2_win')
test.head()

clf = LassoCV(cv=5)

# Set a minimum threshold of 0.25
sfm = SelectFromModel(clf, threshold='mean')
sfm.fit(X, y)
selectedFeatures = sfm.transform(X)

test_selected = sfm.transform(X_test)

### Modelling

svm_clf = svm.SVC(cache_size=7000)
params = {
    'kernel': ['linear','rbf','sigmoid'],
    'C' : np.arange(0.0, 0.5, 0.001),
    'tol' : np.arange(0.0, 0.1, 0.0001),
}

random_serach = RandomizedSearchCV(svm_clf,params,cv=10,random_state=10,refit = True)

start = time()
random_serach.fit(X,y)
print("RandomizedSearchCV took %.2f seconds." % ((time() - start)))

utils.report(random_serach.cv_results_,n_top=10)

SVC_final=svm.SVC(**random_serach.best_params_,probability=True)
SVC_final.fit(X,y)

import pickle
pickle.dump(SVC_final,open("models/SVC_model.sav",'wb'))

y_pred = SVC_final.predict(X_test)
y_pred_prob = SVC_final.predict_proba(X_test)[:, 1]

print("SVM Acc Score: {0:.3%}".format(accuracy_score(y_test,y_pred)))
