==> Bowl_Game_Predictions/2018/3-Predictions.ipynb <==
### Things to be improved in the future 

1) When using raw stats like this, need to think about different paces of play.   
2) Maybe isolate non-conference and conference games (add a feature to mark this)  
3) Develop and use an EPA or WPA model in some way   
4) Conference adjusted stats. 

import pandas as pd
import re

def NumberFromStrings(string):
    numberfromstring = re.findall('\d+',string)
    return numberfromstring

clean_games = pd.read_csv("data/final_data/team_games_data.csv")
clean_games['Team1'] = clean_games['Team1'].str.strip()
clean_games['year'] = clean_games['game_id'].apply(NumberFromStrings).apply(lambda x: x[0])
clean_games_18 = clean_games[clean_games['year'] == '2018'].drop('year',axis=1)
clean_games = clean_games.drop('year',axis=1)
clean_games = clean_games.convert_objects(convert_numeric=True)
clean_games_18 = clean_games_18.convert_objects(convert_numeric=True)

## Load bowl games

bowl_sched = pd.read_csv("data/bowl_sched.csv")

def avg_calcs_function(team_name,team2_cols=False):
    team_df_subset = clean_games_18[clean_games_18['Team1'] == team_name]
    if(len(team_df_subset)==0):
        team_df_subset = clean_games[clean_games['Team1'] == team_name]
    try:
        row = team_df_subset.sample(frac=0.4).drop(['game_id','Team1'],axis=1).mean()
    except: 
        team_df_subset = clean_games 
        row = team_df_subset.sample(frac=0.4).drop(['game_id','Team1'],axis=1).mean()
    if(team2_cols):
        t2_cols_rename = [word.replace('1','2') for word in row.index.tolist()]
        row.index = t2_cols_rename
    return(row)

import pickle
import utils
from sklearn.externals import joblib
scale = joblib.load("models/scaler.sav")
RF_model = pickle.load(open('models/RF_model.sav', 'rb'))
log_model = pickle.load(open('models/Log_model.sav','rb'))
svm_model = pickle.load(open('models/SVC_model.sav','rb'))

def one_time_pred(bowl_df,mod,scale):
    bowl_stats_team1 = bowl_df['Team1'].apply(avg_calcs_function)
    bowl_stats_team2 = bowl_df['Team2'].apply(avg_calcs_function,team2_cols=True)
    bowl_stats_df = pd.concat([bowl_stats_team1,bowl_stats_team2],axis=1)
    order_cols = ['Team1_first_downs', 'Team1_passing_first_downs',
       'Team1_rushing_first_downs', 'Team1_rushing_yds',
       'Team1_rushing_attempts', 'Team1_passing_attempts',
       'Team1_passing_completions', 'Team1_passing_interceptions',
       'Team1_avg_per_pass', 'Team1_pass_yds', 'Team1_yds_per_play',
       'Team1_tot_offense', 'Team1_fumbles_number', 'Team1_fumbles_lost',
       'Team1_pen_number', 'Team1_pen_yds', 'Team1_pun_number',
       'Team1_pun_yds', 'Team1_pun_re_number', 'Team1_pun_re_yds',
       'Team1_kick_re_number', 'Team1_kick_re_yds', 'Team1_int_re_yds',
       'Team1_third_down_attempts', 'Team1_third_down_suc',
       'Team1_fourth_down_attempts', 'Team1_fourth_down_suc',
       'Team2_first_downs', 'Team2_passing_first_downs',
       'Team2_rushing_first_downs', 'Team2_rushing_yds',
       'Team2_rushing_attempts', 'Team2_passing_attempts',
       'Team2_passing_completions', 'Team2_passing_interceptions',
       'Team2_avg_per_pass', 'Team2_pass_yds', 'Team2_yds_per_play',
       'Team2_tot_offense', 'Team2_fumbles_number', 'Team2_fumbles_lost',
       'Team2_pen_number', 'Team2_pen_yds', 'Team2_pun_number',
       'Team2_pun_yds', 'Team2_pun_re_number', 'Team2_pun_re_yds',
       'Team2_kick_re_number', 'Team2_kick_re_yds', 'Team2_int_re_yds',
       'Team2_third_down_attempts', 'Team2_third_down_suc',
       'Team2_fourth_down_attempts', 'Team2_fourth_down_suc',
       'Team1_third_down_conv_pct', 'Team2_third_down_conv_pct',
       'Team1_fourth_down_conv_pct', 'Team2_fourth_down_conv_pct',
       'Team1_turnover_margin', 'Team2_turnover_margin']
    bowl_dfs_clean = bowl_stats_df[order_cols]
    bowl_dfs_clean['Team2_win'] = 0
    X,y,c_names = utils.setup_data(bowl_dfs_clean,scale,'Team2_win',names=True)
    res = mod.predict(X)
    res = res.tolist()
    return(res)

def monte_carlo_preds(mod,n_sims=10):
    res = {}
    for i in range(n_sims):
        res[i] = one_time_pred(bowl_sched,mod,scale)
    res_df = pd.DataFrame(res).sum(axis=1)/n_sims
    return(res_df)

bowl_sched['Team2_win_RF']  = monte_carlo_preds(10000,RF_model,scale)

bowl_sched.sort_values('Team2_win_RF',ascending=False)

log_model.n_jobs=-1

bowl_sched['Team2_win_log'] = monte_carlo_preds(log_model,n_sims=1000)

bowl_sched['Team2_win_SVM'] = monte_carlo_preds(svm_model,n_sims=1000)

bowl_sched.to_csv('data/final_data/predictions.csv')

import numpy as np

np.abs

bowl_sched['conf_RF'] = np.abs(bowl_sched['Team2_win_RF'] - 0.5)
bowl_sched['conf_log'] = np.abs(bowl_sched['Team2_win_log']-0.5)
bowl_sched['conf_svm'] = np.abs(bowl_sched['Team2_win_SVM']-0.5)

bowl_sched.sort_values('conf_svm',ascending=False)
